requirements.txt
flask==3.0.0
torch==2.1.0
torchvision==0.16.0
annoy==1.17.3
Pillow==9.5.0
requests==2.31.0
numpy==1.26.0
flask-cors==4.0.0
tqdm==4.66.1